Refactor: move the writer context files to spark-bigquery-connector-common
