+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/write/context/AvroIntermediateRecordWriter.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.spark.bigquery.v2.context;
+package com.google.cloud.spark.bigquery.write.context;
 
 import java.io.IOException;
 import java.io.OutputStream;
