+++ spark-bigquery-connector-common/src/main/java/com/google/cloud/spark/bigquery/write/context/DataWriterContextFactory.java
@@ -13,8 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.spark.bigquery.v2.context;
+package com.google.cloud.spark.bigquery.write.context;
 
 import java.io.Serializable;
 
-public interface WriterCommitMessageContext extends Serializable {}
+public interface DataWriterContextFactory<T> extends Serializable {
+
+  DataWriterContext<T> createDataWriterContext(int partitionId, long taskId, long epochId);
+}
