+++ spark-bigquery-dsv2/spark-2.4-bigquery/src/main/java/com/google/cloud/spark/bigquery/v2/BigQueryDataSourceV2.java
@@ -18,7 +18,7 @@ package com.google.cloud.spark.bigquery.v2;
 import com.google.cloud.spark.bigquery.InjectorFactory;
 import com.google.cloud.spark.bigquery.v2.context.BigQueryDataSourceReaderContext;
 import com.google.cloud.spark.bigquery.v2.context.BigQueryDataSourceReaderModule;
-import com.google.cloud.spark.bigquery.v2.context.DataSourceWriterContext;
+import com.google.cloud.spark.bigquery.write.context.DataSourceWriterContext;
 import com.google.inject.Injector;
 import java.util.Optional;
 import org.apache.spark.sql.SaveMode;
