+++ spark-bigquery-dsv2/spark-2.4-bigquery/src/main/java/com/google/cloud/spark/bigquery/v2/Spark24DataWriterFactory.java
@@ -15,7 +15,7 @@
  */
 package com.google.cloud.spark.bigquery.v2;
 
-import com.google.cloud.spark.bigquery.v2.context.DataWriterContextFactory;
+import com.google.cloud.spark.bigquery.write.context.DataWriterContextFactory;
 import org.apache.spark.sql.catalyst.InternalRow;
 import org.apache.spark.sql.sources.v2.writer.DataWriter;
 import org.apache.spark.sql.sources.v2.writer.DataWriterFactory;
