+++ spark-bigquery-dsv2/spark-3.1-bigquery/src/main/java/com/google/cloud/spark/bigquery/v2/BigQueryBatchWrite.java
@@ -15,8 +15,8 @@
  */
 package com.google.cloud.spark.bigquery.v2;
 
-import com.google.cloud.spark.bigquery.v2.context.DataSourceWriterContext;
-import com.google.cloud.spark.bigquery.v2.context.WriterCommitMessageContext;
+import com.google.cloud.spark.bigquery.write.context.DataSourceWriterContext;
+import com.google.cloud.spark.bigquery.write.context.WriterCommitMessageContext;
 import java.util.stream.Stream;
 import org.apache.spark.sql.connector.write.BatchWrite;
 import org.apache.spark.sql.connector.write.DataWriterFactory;
