+++ spark-bigquery-dsv2/spark-3.1-bigquery/src/main/java/com/google/cloud/spark/bigquery/v2/BigQueryWriteBuilder.java
@@ -15,7 +15,7 @@
  */
 package com.google.cloud.spark.bigquery.v2;
 
-import com.google.cloud.spark.bigquery.v2.context.DataSourceWriterContext;
+import com.google.cloud.spark.bigquery.write.context.DataSourceWriterContext;
 import com.google.inject.Injector;
 import java.util.Optional;
 import org.apache.spark.sql.SaveMode;
