+++ spark-bigquery-dsv2/spark-bigquery-dsv2-common/pom.xml
@@ -28,17 +28,6 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
-            <groupId>org.apache.beam</groupId>
-            <artifactId>beam-sdks-java-io-hadoop-common</artifactId>
-            <version>2.29.0</version>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.apache.beam</groupId>
-                    <artifactId>beam-sdks-java-core</artifactId>
-                </exclusion>
-            </exclusions>
-        </dependency>
-        <dependency>
             <groupId>org.apache.spark</groupId>
             <artifactId>spark-mllib_2.12</artifactId>
             <version>${spark.version}</version>
@@ -69,32 +58,36 @@
             <scope>test</scope>
         </dependency>
     </dependencies>
-    <build><plugins><plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-enforcer-plugin</artifactId>
-        <dependencies>
-            <dependency>
-                <groupId>de.skuzzle.enforcer</groupId>
-                <artifactId>restrict-imports-enforcer-rule</artifactId>
-                <version>2.0.0</version>
-            </dependency>
-        </dependencies>
-        <executions>
-            <execution>
-                <id>check-spark-imports</id> <!-- put an explanatory ID here -->
-                <phase>process-sources</phase>
-                <goals>
-                    <goal>enforce</goal>
-                </goals>
-                <configuration>
-                    <rules>
-                        <RestrictImports>
-                            <reason>Don't import Spark DataSource V2 classes</reason>
-                            <bannedImport>org.apache.spark.sql.sources.v2.**</bannedImport>
-                        </RestrictImports>
-                    </rules>
-                </configuration>
-            </execution>
-        </executions>
-    </plugin></plugins></build>
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-enforcer-plugin</artifactId>
+                <dependencies>
+                    <dependency>
+                        <groupId>de.skuzzle.enforcer</groupId>
+                        <artifactId>restrict-imports-enforcer-rule</artifactId>
+                        <version>2.0.0</version>
+                    </dependency>
+                </dependencies>
+                <executions>
+                    <execution>
+                        <id>check-spark-imports</id> <!-- put an explanatory ID here -->
+                        <phase>process-sources</phase>
+                        <goals>
+                            <goal>enforce</goal>
+                        </goals>
+                        <configuration>
+                            <rules>
+                                <RestrictImports>
+                                    <reason>Don't import Spark DataSource V2 classes</reason>
+                                    <bannedImport>org.apache.spark.sql.sources.v2.**</bannedImport>
+                                </RestrictImports>
+                            </rules>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+        </plugins>
+    </build>
 </project>
